@@ -55,27 +55,27 @@
     <properties>
         <app.version>${project.version}</app.version>
         <commons-collections4.version>4.4</commons-collections4.version>
-        <commons-io.version>2.13.0</commons-io.version>
-        <java.version>17</java.version>
-        <jdk.source>17</jdk.source>
-        <jdk.target>17</jdk.target>
-        <google-api-services-admin-directory.version>directory_v1-rev20230814-2.0.0</google-api-services-admin-directory.version>
-        <google-api-services-drive.version>v3-rev20230815-2.0.0</google-api-services-drive.version>
-        <google-api-services-groupssettings.version>v1-rev20210624-2.0.0</google-api-services-groupssettings.version>
-        <google-auth-library-oauth2-http.version>1.19.0</google-auth-library-oauth2-http.version>
-        <jquery.version>3.7.0</jquery.version>
-        <lms-canvas-rivet.version>5.2.5.2</lms-canvas-rivet.version>
-        <lms-embedded-services.version>5.2.19</lms-embedded-services.version>
-        <lms-team-spring-boot-it12>4.8</lms-team-spring-boot-it12>
-        <springdoc-openapi-ui.version>1.7.0</springdoc-openapi-ui.version>
-        <spring-cloud-starter-parent.version>2021.0.8</spring-cloud-starter-parent.version>
-        <webjars-locator.version>0.47</webjars-locator.version>
+        <commons-io.version>2.16.1</commons-io.version>
+        <java.version>21</java.version>
+        <jdk.source>21</jdk.source>
+        <jdk.target>21</jdk.target>
+        <google-api-services-admin-directory.version>directory_v1-rev20240903-2.0.0</google-api-services-admin-directory.version>
+        <google-api-services-drive.version>v3-rev20240903-2.0.0</google-api-services-drive.version>
+        <google-api-services-groupssettings.version>v1-rev20220614-2.0.0</google-api-services-groupssettings.version>
+        <google-auth-library-oauth2-http.version>1.25.0</google-auth-library-oauth2-http.version>
+        <jquery.version>3.7.1</jquery.version>
+        <lms-canvas-rivet.version>6.2.8.1_1</lms-canvas-rivet.version>
+        <lms-embedded-services.version>6.0.2</lms-embedded-services.version>
+        <lms-team-spring-boot-it12>6.5.0</lms-team-spring-boot-it12>
+        <springdoc-openapi-ui.version>2.6.0</springdoc-openapi-ui.version>
+        <spring-cloud-starter-parent.version>2023.0.3</spring-cloud-starter-parent.version>
+        <webjars-locator.version>0.52</webjars-locator.version>
 
-        <git-commit-id-maven-plugin.version>6.0.0</git-commit-id-maven-plugin.version>
-        <hibernate5-ddl-maven-plugin.version>2.5.0</hibernate5-ddl-maven-plugin.version>
-        <plugins.compiler.version>3.11.0</plugins.compiler.version>
-        <plugins.license.version>2.2.0</plugins.license.version>
-        <plugins.release.version>3.0.1</plugins.release.version>
+        <git-commit-id-maven-plugin.version>9.0.1</git-commit-id-maven-plugin.version>
+        <hibernate6-ddl-maven-plugin.version>1.0.2</hibernate6-ddl-maven-plugin.version>
+        <plugins.compiler.version>3.13.0</plugins.compiler.version>
+        <plugins.license.version>2.4.0</plugins.license.version>
+        <plugins.release.version>3.1.1</plugins.release.version>
     </properties>
 
     <dependencyManagement>
