@@ -38,108 +38,89 @@ import edu.iu.uits.lms.common.it12logging.RestSecurityLoggingConfig;
 import edu.iu.uits.lms.common.oauth.CustomJwtAuthenticationConverter;
 import edu.iu.uits.lms.lti.repository.DefaultInstructorRoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.security.SecurityProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.access.expression.WebExpressionAuthorizationManager;
+import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
 import org.springframework.security.web.header.writers.ReferrerPolicyHeaderWriter;
 import uk.ac.ox.ctl.lti13.Lti13Configurer;
 
-import static edu.iu.uits.lms.lti.LTIConstants.BASE_USER_ROLE;
+import static edu.iu.uits.lms.lti.LTIConstants.BASE_USER_AUTHORITY;
 import static edu.iu.uits.lms.lti.LTIConstants.WELL_KNOWN_ALL;
 
 @Configuration
+@EnableWebSecurity
 public class SecurityConfig {
 
-    @Configuration
-    @Order(SecurityProperties.BASIC_AUTH_ORDER - 5)
-    public static class GctRestSecurityConfigurationAdapter extends WebSecurityConfigurerAdapter {
+    @Autowired
+    private DefaultInstructorRoleRepository defaultInstructorRoleRepository;
 
-        @Override
-        public void configure(HttpSecurity http) throws Exception {
-            http.cors().and()
-                  .requestMatchers().antMatchers("/rest/**", "/api/**")
-                  .and()
-                  .authorizeRequests()
-                  .antMatchers("/rest/**")
-                  .access("hasAuthority('SCOPE_lms:rest') and hasAuthority('ROLE_LMS_REST_ADMINS')")
-                  .antMatchers("/api/**").permitAll()
-                  .and()
-                  .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                  .and()
-                  .oauth2ResourceServer()
-                  .jwt().jwtAuthenticationConverter(new CustomJwtAuthenticationConverter());
-
-            http.apply(new RestSecurityLoggingConfig());
-        }
+    @Order(5)
+    @Bean
+    public SecurityFilterChain restFilterChain(HttpSecurity http) throws Exception {
+        http.cors(Customizer.withDefaults())
+                .securityMatcher("/rest/**", "/api/**")
+                .authorizeHttpRequests((authz) -> authz
+                        .requestMatchers("/rest/**")
+                        .access(new WebExpressionAuthorizationManager("hasAuthority('SCOPE_lms:rest') and hasAuthority('ROLE_LMS_REST_ADMINS')"))
+                        .requestMatchers("/api/**").permitAll()
+                )
+                .sessionManagement((session) -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+                .oauth2ResourceServer(oauth -> oauth
+                        .jwt(jwt ->
+                                jwt.jwtAuthenticationConverter(new CustomJwtAuthenticationConverter())))
+                .with(new RestSecurityLoggingConfig(), log -> {});
+        return http.build();
     }
 
-    @Configuration
-    @Order(SecurityProperties.BASIC_AUTH_ORDER - 4)
-    public static class GctWebSecurityConfigurerAdapter extends WebSecurityConfigurerAdapter {
+    @Order(6)
+    @Bean
+    public SecurityFilterChain appFilterChain(HttpSecurity http) throws Exception {
+        http.securityMatcher(WELL_KNOWN_ALL, "/error", "/app/**")
+                .authorizeHttpRequests((authz) -> authz
+                        .requestMatchers(WELL_KNOWN_ALL, "/error").permitAll()
+                        .requestMatchers("/**").hasAuthority(BASE_USER_AUTHORITY)
+                        .withObjectPostProcessor(new LmsFilterSecurityInterceptorObjectPostProcessor())
+                )
+                .headers(headers -> headers
+                        .contentSecurityPolicy(csp ->
+                                csp.policyDirectives("style-src 'self' 'unsafe-inline'; form-action 'self'; frame-ancestors 'self' https://*.instructure.com"))
+                        .referrerPolicy(referrer -> referrer
+                                .policy(ReferrerPolicyHeaderWriter.ReferrerPolicy.SAME_ORIGIN))
+                );
 
-        @Autowired
-        private DefaultInstructorRoleRepository defaultInstructorRoleRepository;
-
-        @Override
-        protected void configure(HttpSecurity http) throws Exception {
-            http
-                  .requestMatchers()
-                  .antMatchers(WELL_KNOWN_ALL, "/error", "/app/**")
-                  .and()
-                  .authorizeRequests()
-                  .antMatchers(WELL_KNOWN_ALL, "/error").permitAll()
-                  .antMatchers("/**").hasRole(BASE_USER_ROLE)
-                  .withObjectPostProcessor(new LmsFilterSecurityInterceptorObjectPostProcessor())
-                    .and()
-                    .headers()
-                    .contentSecurityPolicy("style-src 'self' 'unsafe-inline'; form-action 'self'; frame-ancestors 'self' https://*.instructure.com")
-                    .and()
-                    .referrerPolicy(referrer -> referrer
-                            .policy(ReferrerPolicyHeaderWriter.ReferrerPolicy.SAME_ORIGIN));
-
-            //Setup the LTI handshake
-            Lti13Configurer lti13Configurer = new Lti13Configurer()
-                  .grantedAuthoritiesMapper(new CustomRoleMapper(defaultInstructorRoleRepository));
-
-            http.apply(lti13Configurer);
-
-            //Fallback for everything else
-            http.requestMatchers().antMatchers("/**")
-                  .and()
-                  .authorizeRequests()
-                  .anyRequest().authenticated()
-                    .withObjectPostProcessor(new LmsFilterSecurityInterceptorObjectPostProcessor())
-                    .and()
-                    .headers()
-                    .contentSecurityPolicy("style-src 'self' 'unsafe-inline'; form-action 'self'; frame-ancestors 'self' https://*.instructure.com")
-                    .and()
-                    .referrerPolicy(referrer -> referrer
-                            .policy(ReferrerPolicyHeaderWriter.ReferrerPolicy.SAME_ORIGIN));
-        }
-
-        @Override
-        public void configure(WebSecurity web) throws Exception {
-            // ignore everything except paths specified
-            web.ignoring().antMatchers("/app/jsrivet/**", "/app/webjars/**", "/actuator/**", "/app/css/**", "/app/js/**", "/favicon.ico");
-        }
-
+        return http.build();
     }
 
+    @Bean
+    public WebSecurityCustomizer webSecurityCustomizer() {
+        return (web) -> web.ignoring().requestMatchers("/app/jsrivet/**", "/app/webjars/**", "/app/css/**", "/app/js/**", "/favicon.ico");
+    }
 
-    @Configuration
-    @Order(SecurityProperties.BASIC_AUTH_ORDER - 2)
-    public static class GctCatchAllSecurityConfig extends WebSecurityConfigurerAdapter {
+    @Bean
+    public SecurityFilterChain catchallFilterChain(HttpSecurity http) throws Exception {
+        //Setup the LTI handshake
+        http.with(new Lti13Configurer(), lti ->
+                lti.setSecurityContextRepository(new HttpSessionSecurityContextRepository())
+                        .grantedAuthoritiesMapper(new CustomRoleMapper(defaultInstructorRoleRepository)));
 
-        @Override
-        public void configure(HttpSecurity http) throws Exception {
-            http.requestMatchers().antMatchers("/**")
-                  .and()
-                  .authorizeRequests()
-                  .anyRequest().authenticated();
-        }
+        http.securityMatcher("/**")
+                .authorizeHttpRequests((authz) -> authz.anyRequest().authenticated()
+                        .withObjectPostProcessor(new LmsFilterSecurityInterceptorObjectPostProcessor()))
+                .headers(headers -> headers
+                        .contentSecurityPolicy(csp ->
+                                csp.policyDirectives("style-src 'self' 'unsafe-inline'; form-action 'self'; frame-ancestors 'self' https://*.instructure.com"))
+                        .referrerPolicy(referrer -> referrer
+                                .policy(ReferrerPolicyHeaderWriter.ReferrerPolicy.SAME_ORIGIN))
+                );
+
+        return http.build();
     }
 }
